+++ agent/src/main/java/com/cmclinnovations/agent/service/core/QueryTemplateService.java
@@ -73,16 +73,17 @@ public class QueryTemplateService {
   /**
    * Generates a DELETE SPARQL query from the inputs.
    * 
-   * @param resourceID The target resource identifier for the instance.
-   * @param targetId   The target instance IRI.
-   * @param branchName The branch name to filter (can be null).
+   * @param resourceID  The target resource identifier for the instance.
+   * @param targetId    The target instance IRI.
+   * @param branchName  The branch name to filter (can be null).
+   * @param optVarNames Set of names of optional variables.
    */
-  public String genDeleteQuery(String resourceID, String targetId, String branchName) {
+  public String genDeleteQuery(String resourceID, String targetId, String branchName, Set<String> optVarNames) {
     LOGGER.debug("Generating the DELETE query with branchName = {}", branchName);
     // Retrieve the instantiation JSON schema
     ObjectNode addJsonSchema = this.getJsonLDResource(resourceID).deepCopy();
     return this.deleteQueryTemplateFactory
-        .write(new QueryTemplateFactoryParameters(addJsonSchema, targetId, branchName));
+        .write(new QueryTemplateFactoryParameters(addJsonSchema, targetId, branchName, optVarNames));
   }
 
   /**