+++ agent/src/main/java/com/cmclinnovations/agent/utils/QueryResource.java
@@ -107,6 +107,9 @@ public class QueryResource {
 
     public static final String ADD_BRANCH_KEY = "branch_add";
     public static final String DELETE_BRANCH_KEY = "branch_delete";
+    public static final String FIXED_DATE_DATE_KEY = "entry_date";
+    public static final String FIXED_DATE_SCHEDULE_KEY = "schedule entry";
+    public static final String FIXED_DATE_SCHEDULE_DATE_KEY = "schedule entry date";
 
     // Private constructor to prevent instantiation
     private QueryResource() {
@@ -306,6 +309,10 @@ public class QueryResource {
                                 LocalisationTranslator
                                         .getMessage(LocalisationResource.LABEL_ALTERNATE_DAY_SERVICE_KEY))) {
                             return LifecycleResource.RECURRENCE_ALT_DAY_TASK_STRING;
+                        } else if (scheduleTypeContent.equals(
+                                LocalisationTranslator
+                                        .getMessage(LocalisationResource.LABEL_FIXED_DATE_SERVICE_KEY))) {
+                            return LifecycleResource.RECURRENCE_FIXED_DATE_TASK_STRING;
                         }
                         return scheduleType;
                     }).collect(Collectors.toSet());